testtask/requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==3.2.0
packaging==23.2
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
texttable==1.7.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
websocket-client==0.59.0
